% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BsOWT.R
\name{BsOWT}
\alias{BsOWT}
\title{Generate B-spline wavelet transform matrix}
\usage{
BsOWT(x = NULL, type = NULL)
}
\arguments{
\item{x}{Vector corresponding to a quantitative explanatory variable}

\item{type}{Character string specifying the type of B-spline wavelet, being one of `haar`, `linear` or `cubic`}
}
\value{
List of output. `W.tilde` contains the wavelet transform matrix
}
\description{
This function generates the B-spline wavelet transform matrix. It requires
the user to specify a quantitative explanatory variable, along with the type
of B-spline wavelet (one of `haar`, `linear` or `cubic`). Additionally,
optional options include whether to calculate the corresponding wavelet
coefficients for a given response variable.
}
